WIP: create a new build.js that uses just
